+++ app/project/page.tsx
@@ -32,7 +32,7 @@ const ProjectListing = () => {
   }, [searchTerm, filter]);
 
   return (
-    <div className="p-4 sm:p-6 md:p-10 lg:p-14 flex flex-col divide-y-2 divide-slate-700">
+    <div className="p-4 sm:p-6 md:p-10 lg:p-14 !pt-2 flex flex-col divide-y-2 divide-slate-700">
       <PageHeader text="Back to Projects" redirectUrl="/#projects" />
       <div className="flex flex-col space-y-8 py-5">
         <div className="flex flex-col">